[Feature] Add InstallTerminalBenchPkg hook for prebuilt sandbox package upload
- Add an InstallTerminalBenchPkg hook that ships a host-built
  `pip install --target=<dir>` tarball into the sandbox and extracts it to
  /tmp/tb_pkg, letting downstream commands import off PYTHONPATH without an
  online pip install; export it from the public package surface.
- RunAgentInstallDeps logs install-deps stdout/stderr through the pipeline
  logger so install failures are diagnosable from the run log.
- Tighten DetachedShellEntry.handshake_timeout_sec from float to int to match
  downstream timeout fields.

+++ xtuner/v1/rl/agent_loop/sandbox_agent_loop/sandbox.py
@@ -551,7 +551,7 @@ class DetachedShellEntry:
         capture: EntryCapture | dict[str, Any],
         monitor: EntryMonitor | dict[str, Any],
         failure: EntryFailurePolicy | dict[str, Any] | None = None,
-        handshake_timeout_sec: float = 60.0,
+        handshake_timeout_sec: int = 60,
     ):
         self.cmd = cmd
         self.name = name
